+++ core/src/lib/gui/mod.rs
@@ -1,2 +1,3 @@
 mod search_teams;
 mod team;
+mod search_champions;
