+++ core/src/lib/gui/search_teams.rs
@@ -1,13 +1,12 @@
-use std::collections::{ HashMap, HashSet };
-
 use serde::{ Deserialize, Serialize };
 use serde_wasm_bindgen::to_value;
 use wasm_bindgen::{ prelude::wasm_bindgen, JsValue, UnwrapThrowExt };
-use crate::lib::data::{ Champion, ChampionId, GameData, TraitId };
+use crate::lib::data::{ ChampionId, GameData };
 
 use crate::console::log;
 use crate::lib::sat::{ build_champion_constraints, SubgraphSolver };
 
+use super::search_champions::{ ChampionFilter, filter_champions };
 use super::team::Team;
 
 #[derive(Debug, Serialize, Deserialize)]
@@ -17,17 +16,8 @@ struct SearchOptions {
     pub debug: Option<bool>,
 }
 
-#[derive(Debug, Serialize, Deserialize)]
-struct ChampionFilter {
-    ids: Option<Vec<ChampionId>>,
-    cost: Option<Vec<u8>>,
-    range: Option<Vec<u8>>,
-    trait_ids: Option<Vec<TraitId>>,
-    uses_ap: Option<bool>,
-}
-
 #[wasm_bindgen]
-pub fn search_teams(options: JsValue) {
+pub fn search_teams(options: JsValue) -> JsValue {
     let options: SearchOptions = serde_wasm_bindgen
         ::from_value(options)
         .unwrap_throw();
@@ -81,73 +71,6 @@ pub fn search_teams(options: JsValue) {
             }
         }
     }
-}
 
-#[wasm_bindgen]
-pub fn search_champions(filter: JsValue) -> JsValue {
-    let filter: ChampionFilter = serde_wasm_bindgen
-        ::from_value(filter)
-        .unwrap_throw();
-
-    let champions = filter_champions(filter);
-
-    to_value(&champions).unwrap()
-}
-
-fn filter_champions(filter: ChampionFilter) -> Vec<ChampionId> {
-    // @TODO: Is there a way to avoid reinit'ing the champion data every call?
-    //        While still exposing this function to JS. Cant wasm-bindgen return closures?
-    let data = GameData::new();
-
-    let mut champions: Vec<Champion> = Vec::from_iter(
-        data.champions.into_values()
-    );
-
-    if let Some(ids) = filter.ids {
-        let ids = HashSet::<ChampionId>::from_iter(ids.into_iter());
-        champions = champions
-            .into_iter()
-            .filter(|c| ids.contains(&c.name))
-            .collect();
-    }
-
-    if let Some(costs) = filter.cost {
-        let costs = HashSet::<u8>::from_iter(costs.into_iter());
-        champions = champions
-            .into_iter()
-            .filter(|c| costs.contains(&c.cost))
-            .collect();
-    }
-
-    if let Some(ranges) = filter.range {
-        let ranges = HashSet::<u8>::from_iter(ranges.into_iter());
-        champions = champions
-            .into_iter()
-            .filter(|c| ranges.contains(&c.range))
-            .collect();
-    }
-
-    if let Some(trait_ids) = filter.trait_ids {
-        let trait_ids = HashSet::<TraitId>::from_iter(
-            trait_ids.into_iter()
-        );
-        champions = champions
-            .into_iter()
-            .filter(|c|
-                c.traits.iter().any(|t| trait_ids.contains(t))
-            )
-            .collect();
-    }
-
-    if let Some(uses_ap) = filter.uses_ap {
-        champions = champions
-            .into_iter()
-            .filter(|c| c.uses_ap == uses_ap)
-            .collect();
-    }
-
-    champions
-        .iter()
-        .map(|c| c.name.clone())
-        .collect()
+    to_value(&results).unwrap()
 }
